refactor. models 폴더 하위로 entities 연관 코드 이동
