+++ src/refactoring/components/coupon/Coupon.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../../types';
+import { Coupon } from '../../models/coupon/types';
 import CartCouponApplied from './CouponApplied';
 import CartCouponSelectBox from './CouponSelectBox';
