+++ src/refactoring/components/coupon/CouponManagement.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../../types';
+import { Coupon } from '../../models/coupon/types';
 interface Props {
   coupons: Coupon[];
   setNewCoupon: (coupon: Coupon) => void;
