+++ src/refactoring/components/product/ProductEdit.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../../types';
+import { Product } from '../../models/product/types';
 
 interface Props {
   editingProduct: Product;
