+++ src/refactoring/components/product/ProductListItem.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../../types';
+import { Product } from '../../models/product/types';
 import { getMaxDiscount } from '../../utils/getMaxDiscount';
 
 interface Props {
