+++ src/refactoring/components/product/ProductToggleListItem.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../../types';
+import { Product } from '../../models/product/types';
 
 interface Props {
   product: Product;
