+++ src/refactoring/models/cart/types.ts
@@ -1,0 +1,13 @@
+import { Coupon } from '../coupon/types';
+import { Product } from '../product/types';
+
+export interface CartItem {
+  product: Product;
+  quantity: number;
+}
+
+export interface TotalDiscountProps {
+  cart: CartItem[];
+  coupon: Coupon | null;
+  totalBeforeDiscount: number;
+}
