+++ src/refactoring/models/product/types.ts
@@ -1,0 +1,12 @@
+export interface Product {
+  id: string;
+  name: string;
+  price: number;
+  stock: number;
+  discounts: Discount[];
+}
+
+export interface Discount {
+  quantity: number;
+  rate: number;
+}
