+++ src/refactoring/utils/getMaxDiscount.ts
@@ -1,4 +1,4 @@
-import { Discount } from '../../types';
+import { Discount } from '../models/product/types';
 
 export const getMaxDiscount = (discounts: Discount[]) =>
   discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
